+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/assets/AssetRepository.java
@@ -17,4 +17,7 @@ public interface AssetRepository extends JpaRepository<Asset, Long> {
 
   @Transactional
   void deleteByExternalId(String gameId);
+
+  @Transactional
+  void deleteByUuid(String gameId);
 }